more bits and pieces
go through these again carefully, and remove the #pragma's.

+++ sys/sys/mount.h
@@ -452,7 +452,7 @@ extern long buflowpages, bufhighpages, bufbackpages;
 #define BUFPAGES_INACT (((bcstats.numcleanpages - buflowpages) < 0) ? 0 \
     : bcstats.numcleanpages - buflowpages)
 extern int bufcachepercent;
-extern void bufadjust(int);
+extern void bufadjust(long);
 struct uvm_constraint_range;
 extern int bufbackoff(struct uvm_constraint_range*, long);
